add dropCache method
invalidate the whole cache with one command.
this can be useful if you need to invalidate the entire cache when you change certain routes.
think about a main menu and a sub menu: when the use navigates the sub menu, all the sub routes are cached and fetched from the cache. then, when the user navigates the main menu, all the cached routes need to be dropped.